Use i64 for Telegram IDs in the database

+++ migrations/2018-07-14-112340_create_chats/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE chat (
-  telegram_id   INT         PRIMARY KEY,
+  telegram_id   BIGINT      PRIMARY KEY,
   title         TEXT        NOT NULL,
   created_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP,
   updated_at    DATETIME    NOT NULL        DEFAULT CURRENT_TIMESTAMP
