+++ src/models.rs
@@ -2,7 +2,7 @@ use chrono::NaiveDateTime;
 
 #[derive(Queryable)]
 pub struct Chat {
-    pub telegram_id: i32,
+    pub telegram_id: i64,
     pub title: String,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
@@ -10,7 +10,7 @@ pub struct Chat {
 
 #[derive(Queryable)]
 pub struct User {
-    pub telegram_id: i32,
+    pub telegram_id: i64,
     pub name: String,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
@@ -18,8 +18,8 @@ pub struct User {
 
 #[derive(Queryable)]
 pub struct ChatUserStats {
-    pub chat_id: i32,
-    pub user_id: i32,
+    pub chat_id: i64,
+    pub user_id: i64,
     pub messages: u32,
     pub created_at: NaiveDateTime,
     pub updated_at: NaiveDateTime,
